beta 0.2

+++ src/Services/CompanyService.php
@@ -48,6 +48,13 @@
             if (!is_numeric($data['companyID'])) {
                 throw new \Exception('Company ID may consist digits only.', 403);
             }
+            if (!$this->companyRepository->isEmptyWarehouses($data['companyID'])) {
+                throw new \Exception('Some warehouses have items. Delete is prohibited.', 403);
+            }
+            if (!$this->companyRepository->isPendingTransfers($data['companyID'])) {
+                throw new \Exception('Some transfers are not accepted. Delete is prohibited.', 403);
+            }
+
             return $this->companyRepository->delete($data);
         }
     }
